🚚 [Move] 디렉토리 구조 변경
디렉토리 구조 변경 (전부 src 안으로 넣기)
