+++ src/services/getDiffingResultQuery.js
@@ -1,4 +1,5 @@
 import { useQuery } from "react-query";
+
 import generateApiUri from "../utils/generateURI";
 
 const getDiffingResult = async (
