+++ src/utils/createDiffText.js
