+++ src/utils/generateURI.js
